Fix offset of radial distance index in EDS
gl_FragCoord has half-integer values, while the trapezoid rule is
implemented in a form that expects integer input indices.

+++ shaders/compute-eclipsed-double-scattering.frag
@@ -78,7 +78,7 @@ void main()
                                                                          viewRayIntersectsGround);
 
     CONST int directionIndex=int(gl_FragCoord.x);
-    CONST float radialDistIndex=gl_FragCoord.y;
+    CONST int radialDistIndex=int(gl_FragCoord.y);
 
     // Using midpoint rule for quadrature
     CONST float dl=radialIntegrInterval/radialIntegrationPoints;
